Created AutocompleteSearch off WebsiteBuilder, Added Loading Animation
